+++ packages/suite-base/src/components/MessagePathSyntax/MessagePathInput.style.ts
@@ -1,0 +1,8 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2025 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import { makeStyles } from "tss-react/mui";
+
+export const useStyles = makeStyles()({
+  root: { flexGrow: 1 },
+});
